Purpose: Fix bug Description:     - Added COMP_CODE_IMCOMP to comp_coder_t with value 12 (same as       COMP_IMCOMP) so that IMCOMP compression can be detected     - Modified various library functions to include COMP_CODE_IMCOMP       so that IMCOMP compression cannot be used in writing but only       reading     - Replaced grgetcomptype with GRgetcomptype in the library and test       (Will do the same in tools)     - Removed many warnings and added notes about some other warnings       so that they can be reviewed again later.
Platforms Tested:
    Linux/32 2.6 (jam)
    FreeBSD 8.2 (freedom)
    Linux/64 2.6 (amani)

+++ hdf/src/hdfi.h
@@ -1436,7 +1436,7 @@ correctly.
         (i) |= (uint16)(*(p) & 0xff); (p)++; }
 
 #   define INT32DECODE(p, i) \
-{ (i) = ((*(p) & 0x80) ? ~0xffffffff : 0x00) | ((int32)(*(p) & 0xff) << 24); (p)++; \
+{ (i) = (int32)(((int32)*(p) & 0x80) ? ~0xffffffff : 0x00) | ((int32)(*(p) & 0xff) << 24); (p)++; \
         (i) |= ((int32)(*(p) & 0xff) << 16); (p)++; \
         (i) |= ((int32)(*(p) & 0xff) << 8); (p)++; \
         (i) |= (*(p) & 0xff); (p)++; }
